#435 Change place to define type validator
